+++ src/test/java/com/atlassian/mail/server/OtherTestSmtpMailServer.java
@@ -1,7 +1,6 @@
-package atlassian.mail.server;
+package com.atlassian.mail.server;
 
 import com.atlassian.mail.MailProtocol;
-import com.atlassian.mail.server.SMTPMailServer;
 
 public interface OtherTestSmtpMailServer extends DefaultTestMailServer, SMTPMailServer {
