+++ src/test/java/com/atlassian/mail/server/OtherTestSmtpMailServerImpl.java
@@ -1,10 +1,10 @@
-package atlassian.mail.server;
+package com.atlassian.mail.server;
 
 import com.atlassian.mail.server.impl.SMTPMailServerImpl;
 
-public class DefaultTestSmtpMailServerImpl extends SMTPMailServerImpl implements DefaultTestSmtpMailServer {
+public class OtherTestSmtpMailServerImpl extends SMTPMailServerImpl implements OtherTestSmtpMailServer {
 
-    public DefaultTestSmtpMailServerImpl() {
+    public OtherTestSmtpMailServerImpl() {
         super(
                 null,
                 NAME,
